Filled out a bit of the Window class. Mostly just wrappers to the SDL2 window functions.
Application now sets the logical render size to 1680x1050. The window is
still only 800x600, but now the application assumes rendering is done on
the larger rendering area and scales it all down. WHY did I do this? I
wanted to see it in action. This will be handled in a MUCH more
appropriate manner later.